chore: rename components from scaffold-stylus to scaffold-eth
