+++ packages/nextjs/app/blockexplorer/_components/AddressComponent.tsx
@@ -1,6 +1,6 @@
 import { BackButton } from "./BackButton";
 import { ContractTabs } from "./ContractTabs";
-import { Address, Balance } from "~~/components/scaffold-stylus";
+import { Address, Balance } from "~~/components/scaffold-eth";
 
 export const AddressComponent = ({
   address,
