+++ packages/nextjs/app/blockexplorer/_components/TransactionsTable.tsx
@@ -1,6 +1,6 @@
 import { TransactionHash } from "./TransactionHash";
 import { formatEther } from "viem";
-import { Address } from "~~/components/scaffold-stylus";
+import { Address } from "~~/components/scaffold-eth";
 import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
 import { TransactionWithFunction } from "~~/utils/scaffold-stylus";
 import { TransactionsTableProps } from "~~/utils/scaffold-stylus";
