+++ packages/nextjs/app/debug/_components/contract/utilsDisplay.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { TransactionBase, TransactionReceipt, formatEther, isAddress } from "viem";
-import { Address } from "~~/components/scaffold-stylus";
+import { Address } from "~~/components/scaffold-eth";
 import { replacer } from "~~/utils/scaffold-stylus/common";
 
 type DisplayContent =
