+++ packages/nextjs/components/scaffold-eth/BlockieAvatar.tsx
