+++ packages/nextjs/components/scaffold-eth/Input/Bytes32Input.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from "react";
 import { hexToString, isHex, stringToHex } from "viem";
-import { CommonInputProps, InputBase } from "~~/components/scaffold-stylus";
+import { CommonInputProps, InputBase } from "~~/components/scaffold-eth";
 
 export const Bytes32Input = ({ value, onChange, name, placeholder, disabled }: CommonInputProps) => {
   const convertStringToBytes32 = useCallback(() => {
