+++ packages/nextjs/components/scaffold-eth/RainbowKitCustomConnectButton/AddressQRCodeModal.tsx
@@ -1,6 +1,6 @@
 import { QRCodeSVG } from "qrcode.react";
 import { Address as AddressType } from "viem";
-import { Address } from "~~/components/scaffold-stylus";
+import { Address } from "~~/components/scaffold-eth";
 
 type AddressQRCodeModalProps = {
   address: AddressType;
